requirements.txt
# List your Python package dependencies here
Flask
pymongo
python-dotenv
werkzeug
